connected the hero button & the storytelling button to the modal
