+++ journeywise-app/src/App.jsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import './App.css'
-import SecNoticePage from './pages/SecNoticePage'
 import Index from './navigation/Index'
 import Header from './components/Nav/Header'
