--- journeywise-app/src/pages/SecNoticePage.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import SecNotice from '../components/SecNotice/SecNotice'
-function SecNoticePage() {
-  return <>{<SecNotice />}</>
-}
-
-export default SecNoticePage
